moved msgqueue to network + networ.message to separate file
